+++ src/controllers/book/booksControllers.ts
@@ -2,6 +2,7 @@ import { Response, Request } from 'express';
 import { connection } from '../../database/connection';
 import crypto from 'crypto';
 import { Book, BookProps } from '../../types/booksControllers'
+import checkAuthorization from '../../utils/checkAuthorization';
 
 export default {
     async index(req: Request, res: Response) {
@@ -19,6 +20,7 @@ export default {
 
     async create(req: Request, res: Response) {
         const book = req.body;
+        const { authorization } = req.headers;
         const id = crypto.randomBytes(4).toString('hex');
 
         const values: BookProps['create']['crud']['tableBook'] = book;
@@ -37,7 +39,13 @@ export default {
             cddcdu: values.cddcdu,
         };
 
+
         try {
+            const validateAuthorization = await checkAuthorization(authorization, 2);
+
+            if (!validateAuthorization.status) {
+                throw new Error(validateAuthorization.error);
+            };
             if (values.amount && values.salePrice && values.factoryPrice) {
                 const valuesBookStock = {
                     idBook: id,
@@ -87,8 +95,14 @@ export default {
 
     async Delete(req: Request, res: Response) {
         const { id: idBook } = req.params;
+        const { authorization } = req.headers;
 
         try {
+            const validateAuthorization = await checkAuthorization(authorization, 2);
+
+            if (!validateAuthorization.status) {
+                throw new Error(validateAuthorization.error);
+            };
             await connection('books')
                 .where('id', idBook)
                 .delete();
